+++ docs/diffusers/wandb_callbacks.md
@@ -2,10 +2,10 @@
 
 Callbacks for logging experiment details, configs and generated images for [Diffusers 🧨](https://huggingface.co/docs/diffusers) pipelines to your [Weights & Biases workspace](https://docs.wandb.ai/guides/app/pages/workspaces).
 
-::: wandb_addons.diffusers.stable_diffusion
+::: wandb_addons.diffusers.callbacks.stable_diffusion
 
-::: wandb_addons.diffusers.kandinsky
+::: wandb_addons.diffusers.callbacks.kandinsky
 
-::: wandb_addons.diffusers.base
+::: wandb_addons.diffusers.callbacks.base
 
 ::: wandb_addons.diffusers.wrapper
