+++ wandb_addons/diffusers/__init__.py
@@ -1,5 +1,5 @@
-from .stable_diffusion import StableDiffusionCallback
-from .kandinsky import KandinskyCallback
+from .callbacks import StableDiffusionCallback
+from .callbacks import KandinskyCallback
 from .wrapper import get_wandb_callback
 
 __all__ = ["StableDiffusionCallback", "KandinskyCallback", "get_wandb_callback"]
