+++ wandb_addons/diffusers/callbacks/__init__.py
@@ -1,0 +1,4 @@
+from .stable_diffusion import StableDiffusionCallback
+from .kandinsky import KandinskyCallback
+
+__all__ = ["StableDiffusionCallback", "KandinskyCallback"]
