+++ wandb_addons/diffusers/callbacks/kandinsky.py
@@ -1,17 +1,15 @@
 from typing import Dict, List, Optional, Union
 
 import torch
-import wandb
-from diffusers import StableDiffusionPipeline
-from PIL import Image
+from diffusers import DiffusionPipeline, KandinskyCombinedPipeline, KandinskyPipeline
 
 from .base import BaseDiffusersBaseCallback
 
 
-class StableDiffusionCallback(BaseDiffusersBaseCallback):
+class KandinskyCallback(BaseDiffusersBaseCallback):
     """Callback for [🧨 Diffusers](https://huggingface.co/docs/diffusers/index) logging
     the results of a
-    [`StableDiffusionPipeline`](https://huggingface.co/docs/diffusers/v0.9.0/en/api/pipelines/stable_diffusion#diffusers.StableDiffusionPipeline)
+    [`KandinskyCombinedPipeline`](https://huggingface.co/docs/diffusers/api/pipelines/kandinsky#diffusers.KandinskyCombinedPipeline)
     generation to Weights & Biases.
 
     !!! note "Features:"
@@ -24,19 +22,19 @@ class StableDiffusionCallback(BaseDiffusersBaseCallback):
             runs gracefully.
 
     !!! example "Example usage:"
-        You can fine an example notebook [here](../examples/stable_diffusion).
+        You can fine an example notebook [here](../examples/kandinsky).
 
         ```python
         import torch
-        from diffusers import StableDiffusionPipeline
+        from diffusers import KandinskyCombinedPipeline
 
-        from wandb_addons.diffusers import StableDiffusionCallback
+        from wandb_addons.diffusers import KandinskyCallback
 
 
-        pipeline = StableDiffusionPipeline.from_pretrained(
-            "CompVis/stable-diffusion-v1-4", torch_dtype=torch.float16
+        pipe = KandinskyCombinedPipeline.from_pretrained(
+            "kandinsky-community/kandinsky-2-1", torch_dtype=torch.float16
         )
-        pipeline = pipeline.to("cuda")
+        pipe = pipe.to("cuda")
 
         prompt = [
             "a photograph of an astronaut riding a horse",
@@ -44,13 +42,17 @@ class StableDiffusionCallback(BaseDiffusersBaseCallback):
         ]
         negative_prompt = ["ugly, deformed", "ugly, deformed"]
         num_images_per_prompt = 2
+
         configs = {
-            "eta": 0.0,
-            "guidance_rescale": 0.0,
+            "guidance_scale": 4.0,
+            "height": 512,
+            "width": 512,
+            "prior_guidance_scale": 4.0,
+            "prior_num_inference_steps": 25,
         }
 
         # Create the WandB callback for StableDiffusionPipeline
-        callback = StableDiffusionCallback(
+        callback = KandinskyCallback(
             pipe,
             prompt=prompt,
             negative_prompt=negative_prompt,
@@ -70,8 +72,8 @@ class StableDiffusionCallback(BaseDiffusersBaseCallback):
         ```
 
     Arguments:
-        pipeline (diffusers.StableDiffusionPipeline): The `StableDiffusionPipeline` from
-            `diffusers`.
+        pipeline (Union[DiffusionPipeline, KandinskyCombinedPipeline, KandinskyPipeline]):
+            The `KandinskyCombinedPipeline` or `KandinskyPipeline` from `diffusers`.
         prompt (Union[str, List[str]]): The prompt or prompts to guide the image
             generation.
         wandb_project (Optional[str]): The name of the project where you're sending
@@ -88,28 +90,23 @@ class StableDiffusionCallback(BaseDiffusersBaseCallback):
             usually lead to a higher quality image at the expense of slower inference.
         num_images_per_prompt (Optional[int]): The number of images to generate per
             prompt.
-        guidance_scale (float): Guidance scale as defined in
-            [Classifier-Free Diffusion Guidance](https://arxiv.org/abs/2207.12598).
-            `guidance_scale` is defined as `w` of equation 2. of
-            [Imagen Paper](https://arxiv.org/pdf/2205.11487.pdf). Guidance scale is
-            enabled by setting `guidance_scale > 1`. Higher guidance scale encourages
-            to generate images that are closely linked to the text `prompt`, usually
-            at the expense of lower image quality.
         negative_prompt (Optional[Union[str, List[str]]]): The prompt or prompts not
             to guide the image generation. Ignored when not using guidance
             (i.e., ignored if `guidance_scale` is less than `1`).
         configs (Optional[Dict]): Additional configs for the experiment you want to
-            sync, for example, seed could be a good config to be passed here.
+            sync, for example, for example, `seed` could be a good config to be passed
+            here.
     """
 
     def __init__(
         self,
-        pipeline: StableDiffusionPipeline,
+        pipeline: Union[
+            DiffusionPipeline, KandinskyCombinedPipeline, KandinskyPipeline
+        ],
         prompt: Union[str, List[str]],
         wandb_project: str,
         wandb_entity: Optional[str] = None,
-        guidance_scale: float = 7.5,
-        num_inference_steps: int = 50,
+        num_inference_steps: int = 100,
         num_images_per_prompt: Optional[int] = 1,
         negative_prompt: Optional[Union[str, List[str]]] = None,
         configs: Optional[Dict] = None,
@@ -126,36 +123,13 @@ class StableDiffusionCallback(BaseDiffusersBaseCallback):
             configs,
             **kwargs
         )
-        self.guidance_scale = guidance_scale
-        self.do_classifier_free_guidance = guidance_scale > 1.0
-        wandb.config.update(
-            {
-                "guidance_scale": self.guidance_scale,
-                "do_classifier_free_guidance": self.do_classifier_free_guidance,
-            }
-        )
-
-    def build_wandb_table(self) -> None:
-        super().build_wandb_table()
-        self.table_columns += ["Guidance-Scale", "Do-Classifier-Free-Guidance"]
-
-    def populate_table_row(
-        self, prompt: str, negative_prompt: str, image: Image
-    ) -> None:
-        super().populate_table_row(prompt, negative_prompt, image)
-        self.table_row += [self.guidance_scale, self.do_classifier_free_guidance]
+        self.starting_step = 0
+        self.log_step = num_inference_steps - 1
 
     def generate(self, latents: torch.FloatTensor) -> List:
-        text_embeddings = self.pipeline._encode_prompt(
-            self.prompt,
-            self.pipeline._execution_device,
-            self.num_images_per_prompt,
-            self.do_classifier_free_guidance,
-            self.negative_prompt,
-        )
-        images = self.pipeline.decode_latents(latents)
-        images, _ = self.pipeline.run_safety_checker(
-            images, self.pipeline._execution_device, text_embeddings.dtype
-        )
+        images = self.pipeline.movq.decode(latents, force_not_quantize=True)["sample"]
+        images = images * 0.5 + 0.5
+        images = images.clamp(0, 1)
+        images = images.cpu().permute(0, 2, 3, 1).float().numpy()
         images = self.pipeline.numpy_to_pil(images)
         return images
